moved threshold from separate collection to categories
